(refactor) migrate NetworkStatus to network_iterator
