+++ hummingbot/core/rate_oracle/rate_oracle.py
@@ -9,7 +9,8 @@ from decimal import Decimal
 import aiohttp
 from enum import Enum
 from hummingbot.logger import HummingbotLogger
-from hummingbot.core.network_base import NetworkBase, NetworkStatus
+from hummingbot.core.network_base import NetworkBase
+from hummingbot.core.network_iterator import NetworkStatus
 from hummingbot.core.utils.async_utils import safe_ensure_future
 import hummingbot.client.settings # noqa
 from hummingbot.connector.exchange.binance.binance_utils import convert_from_exchange_trading_pair as \
