+++ hummingbot/data_feed/custom_api_data_feed.py
@@ -2,7 +2,8 @@ import asyncio
 import aiohttp
 import logging
 from typing import Optional
-from hummingbot.core.network_base import NetworkBase, NetworkStatus
+from hummingbot.core.network_base import NetworkBase
+from hummingbot.core.network_iterator import NetworkStatus
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.utils.async_utils import safe_ensure_future
 from decimal import Decimal
